+++ tests/test_math/test_backend_manager.py
@@ -22,6 +22,7 @@ import tensorflow as tf
 from jax import numpy as jnp
 
 from mrmustard import math
+from mrmustard.math.lattice.strategies.displacement import displacement
 
 
 # pylint: disable=too-many-public-methods
@@ -717,3 +718,73 @@ class TestBackendManager:
         probs = np.array([1e-6 for _ in range(300)])
         results = [math.Categorical(probs, "") for _ in range(100)]
         assert len(set(results)) > 1
+
+    def test_displacement(self):
+        r"""
+        Tests the ``displacement`` method.
+        """
+        cutoff = 5
+        alpha = 0.3 + 0.5 * 1j
+        # This data is obtained by using qutip
+        # np.array(displace(40,alpha).data.todense())[0:5,0:5]
+        expected = np.array(
+            [
+                [
+                    0.84366482 + 0.00000000e00j,
+                    -0.25309944 + 4.21832408e-01j,
+                    -0.09544978 - 1.78968334e-01j,
+                    0.06819609 + 3.44424719e-03j,
+                    -0.01109048 + 1.65323865e-02j,
+                ],
+                [
+                    0.25309944 + 4.21832408e-01j,
+                    0.55681878 + 0.00000000e00j,
+                    -0.29708743 + 4.95145724e-01j,
+                    -0.14658716 - 2.74850926e-01j,
+                    0.12479885 + 6.30297236e-03j,
+                ],
+                [
+                    -0.09544978 + 1.78968334e-01j,
+                    0.29708743 + 4.95145724e-01j,
+                    0.31873657 + 0.00000000e00j,
+                    -0.29777767 + 4.96296112e-01j,
+                    -0.18306015 - 3.43237787e-01j,
+                ],
+                [
+                    -0.06819609 + 3.44424719e-03j,
+                    -0.14658716 + 2.74850926e-01j,
+                    0.29777767 + 4.96296112e-01j,
+                    0.12389162 + 1.10385981e-17j,
+                    -0.27646677 + 4.60777945e-01j,
+                ],
+                [
+                    -0.01109048 - 1.65323865e-02j,
+                    -0.12479885 + 6.30297236e-03j,
+                    -0.18306015 + 3.43237787e-01j,
+                    0.27646677 + 4.60777945e-01j,
+                    -0.03277289 + 1.88440656e-17j,
+                ],
+            ]
+        )
+        D = math.displacement(math.real(alpha), math.imag(alpha), (cutoff, cutoff))
+        assert np.allclose(D, expected, atol=1e-5, rtol=0)
+        D_identity = math.displacement(0, 0, (cutoff, cutoff))
+        assert np.allclose(D_identity, np.eye(cutoff), atol=1e-5, rtol=0)
+
+    def test_beamsplitter(self):
+        r"""
+        Tests the ``beamsplitter`` method.
+        """
+        pass
+
+    def test_squeezed(self):
+        r"""
+        Tests the ``squeezed`` method.
+        """
+        pass
+
+    def test_squeezer(self):
+        r"""
+        Tests the ``squeezer`` method.
+        """
+        pass
